backend/requirements.txt
django
djangorestframework
pandas
django-cors-headers
reportlab

